Create de_de.json
This is the most I can do; this are by no means litteral translations as crucible means `tiegel` in german, which would be used in smeltery. As such I used the word for Brewing caldron. For crucible spoon I used the word for mixing spoon instead of `tiegellöffel`, as tiegellöffel would make next to no sense in german